Create DayListItem.js
Component shows us all of the data for a single weekday.

When the user clicks on a particular day, it displays a white background, indicating the user has selected that day.

This component will be used by our DayList component (yet to be created), allowing the user to select from a list of days.