Update InterviewList component
After reviewing the Interviewers array in Index.js, identified that there's no such prop in the Interviewers array passed to this component.

Hence why removed line 16 and replaced the code on line 20, to properly reflect selected users based off elements in the interviewers array and how selected user is defined in Storybook.